chore: rename stream processor to align w/ crud
