+++ product/stream_processor/handlers/product_stream.py
